LPS-105859 Move all SEO logic to SEO display context
